Core parser improvements, hardware register optimization and linting, improving the code style - Refactor memory write function to remove redundant stack write warning - Enhance register parsing to support additional formats and improve error handling - Implement register list parsing for ARM assembly syntax - Update immediate value parsing to support negative values and improve error handling -Add sorting and binary search for hardware register retrieval [O(log n) lookup] - Generate register mapping files with auto-generated warnings and improve structure - Add .clang-format and .editorconfig for consistent code styling

+++ Core/Inc/commands.h
@@ -39,7 +39,7 @@ extern "C"
         const char *mnemonic;                                      ///< Instruction mnemonic (e.g., "MOV", "ADD")
         uint8_t min_operands;                                      ///< Minimum number of operands required
         uint8_t max_operands;                                      ///< Maximum number of operands allowed
-        uint16_t operand_patterns[4];                              ///< Valid operand type patterns per position
+        uint32_t operand_patterns[4];                              ///< Valid operand type patterns per position
         result_t (*validate)(const operand_t *ops, uint8_t count); ///< Custom validation function (optional)
         result_t (*execute)(const operand_t *ops, uint8_t count);  ///< Instruction execution handler
     } instruction_descriptor_t;
